# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Opsy Backend Backend (in GoLang)",
        "title": "Opsy Backend",
        "contact": {
            "name": "Priyanshu Rai"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.0.1"
    },
    "host": "nssdmvtpj2.us-east-1.awsapprunner.com",
    "basePath": "/",
    "paths": {
...
        "/user/delete-account/{id}": {
            "put": {
                "description": "delete account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user authorization"
                ],
                "summary": "delete account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logEntry.DeleteResDto"
                        }
                    }
                }
            }
        },
...
        "/user/forgot-password": {
            "post": {
                "description": "Forgot Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user authorization"
                ],
                "summary": "Forgot Password",
                "parameters": [
                    {
                        "description": "forgot password for user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userAuth.UserForgotPasswordReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userAuth.UserPasswordResDto"
                        }
                    }
                }
            }
        },
...
        "/user/weekly-insights": {
            "get": {
                "description": "Retrieves mental health and physical health data for a specified week",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logEntry"
                ],
                "summary": "Fetch weekly insights data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date of the week (YYYY-MM-DD)",
                        "name": "firstDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date of the week (YYYY-MM-DD)",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logEntry.InsightsResDto"
                        }
                    }
                }
            }
        },
        "/user/yearly-insights": {
            "get": {
                "description": "Retrieves mental health and physical health data for each month of the specified year",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logEntry"
                ],
                "summary": "Fetch monthly insights data for a specific year",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Year (YYYY)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/opsy_backend_handlers_users_logEntry.YearInsightsResDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "logEntry.DeleteResDto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "logEntry.InsightsData": {
            "type": "object",
            "properties": {
                "mentalHealth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logEntry.MentalHealthRes"
                    }
                },
                "physicalHealth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logEntry.PhysicalHealthRes"
                    }
                }
            }
        },
        "logEntry.InsightsResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logEntry.InsightsData"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "logEntry.MentalHealthRes": {
            "type": "object",
            "properties": {
                "avgFeel": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                }
            }
        },
        "logEntry.PhysicalHealthRes": {
            "type": "object",
            "properties": {
                "avgPainLevel": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                }
            }
        },
        "opsy_backend_handlers_users_logEntry.YearInsightsResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/opsy_backend_handlers_users_logEntry.YearlyInsightsRes"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "opsy_backend_handlers_users_logEntry.YearlyInsightsRes": {
            "type": "object",
            "properties": {
                "mentalHealth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/opsy_backend_handlers_users_logEntry.YearlyMentalHealthInsightsResData"
                    }
                },
                "physicalHealth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/opsy_backend_handlers_users_logEntry.YearlyPhysicalHealthInsightsResData"
                    }
                }
            }
        },
        "opsy_backend_handlers_users_logEntry.YearlyMentalHealthInsightsResData": {
            "type": "object",
            "properties": {
                "avgFeel": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "opsy_backend_handlers_users_logEntry.YearlyPhysicalHealthInsightsResData": {
            "type": "object",
            "properties": {
                "avgPainLevel": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                }
            }
        },
...
        "userAuth.UserForgotPasswordReqDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "userAuth.UserPasswordResDto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}